fix: Retirando todos os longs do projeto e substituindo por Integer

+++ app-backend-atelie/src/main/java/com/eseltech/appbackendatelie/DTO/MaterialDTO.java
@@ -18,7 +18,7 @@ public record MaterialDTO(
                 example = "1",
                 accessMode = Schema.AccessMode.READ_ONLY
         )
-        Long id,
+        Integer id,
 
         @NotNull(message = "O ID da empresa é obrigatório")
         @Schema(
@@ -26,7 +26,7 @@ public record MaterialDTO(
                 example = "1",
                 requiredMode = Schema.RequiredMode.REQUIRED
         )
-        Long empresaId,
+        Integer empresaId,
 
         @NotNull(message = "A categoria é obrigatória")
         @Schema(
